Single resource mode supports resource population and depedency
E.g. `aztfy res <vm id>` now might terrafies three resources: vm, managed disk and disk attachement.

+++ internal/test/query/query_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/Azure/aztfy/internal"
 	"github.com/Azure/aztfy/internal/config"
 	"github.com/Azure/aztfy/internal/test"
+	"github.com/Azure/aztfy/internal/utils"
 	"github.com/hashicorp/terraform-exec/tfexec"
 )
 
@@ -18,9 +19,15 @@ func TestQueryMode(t *testing.T) {
 	test.Precheck(t)
 	d := test.NewData()
 	tfexecPath := test.EnsureTF(t)
+
 	provisionDir := t.TempDir()
+	if test.Keep() {
+		provisionDir, _ = os.MkdirTemp("", "")
+		t.Log(provisionDir)
+	}
+
 	os.Chdir(provisionDir)
-	if err := os.WriteFile("main.tf", []byte(fmt.Sprintf(`
+	if err := utils.WriteFileSync("main.tf", []byte(fmt.Sprintf(`
 provider "azurerm" {
   features {
     resource_group {
@@ -60,12 +67,15 @@ resource "azurerm_subnet" "test" {
 	if err := tf.Apply(ctx); err != nil {
 		t.Fatalf("terraform apply failed: %v", err)
 	}
-	defer func() {
-		t.Log("Running: terraform destroy")
-		if err := tf.Destroy(ctx); err != nil {
-			t.Logf("terraform destroy failed: %v", err)
-		}
-	}()
+
+	if !test.Keep() {
+		defer func() {
+			t.Log("Running: terraform destroy")
+			if err := tf.Destroy(ctx); err != nil {
+				t.Logf("terraform destroy failed: %v", err)
+			}
+		}()
+	}
 
 	const delay = time.Minute
 	t.Logf("Sleep for %v to wait for the just created resources be recorded in ARG\n", delay)
